Build rlscope for CUDA 11.0 to support TensorFlow v2.4.0

+++ src/libs/common_util/logging.h
@@ -353,7 +353,7 @@ std::string* MakeCheckOpString(const T1& v1, const T2& v2, const char* exprtext)
   }                                                                       \
   inline std::string* name##Impl(const int v1, const size_t v2,                \
                             const char* exprtext) {                       \
-    if (TF_PREDICT_FALSE(v2 >= std::numeric_limits<int>::max())) {        \
+    if (TF_PREDICT_FALSE(v2 >= static_cast<size_t>(std::numeric_limits<int>::max()))) {        \
       return ::rlscope::internal::MakeCheckOpString(v1, v2, exprtext); \
     }                                                                     \
     const size_t uval = (size_t)((unsigned)v2);                           \
